code optimization
Introduced the State class to represent different states of a node and assigned appropriate colors to each state.
Refactored the code to use a priority queue instead of a regular queue in the BFS algorithm.
Created a reset method in the Node class to reset the state of a node.
Renamed the function draw_grid to draw to avoid conflicting names.
Moved the draw function call outside the main loop in the main function.
Added the __lt__ method to the Node class to allow nodes to be compared in the priority queue.
Updated the make_grid function to use the numpy library for creating the grid array.
Added a reset parameter to the dfs_search function to allow resetting of the grid.
Updated the bfs function to reconstruct the path in reverse order for visualization purposes.
Improved code readability by adding comments and improving variable names.